requirements.txt
streamlit>=1.28.0
sqlalchemy>=2.0.0
google-generativeai>=0.3.0
pandas>=1.5.0
